This will seperate js logic and html to two different components
